% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_fram_fishery_to_crc.R
\name{assign_fram_fishery_to_crc}
\alias{assign_fram_fishery_to_crc}
\title{assign_fram_fishery_to_crc}
\usage{
assign_fram_fishery_to_crc(x, ...)
}
\arguments{
\item{x}{data.frame-like object with character "area_code" column}
}
\value{
object with new (or overwritten) integer column "FisheryID"
}
\description{
Simple wrapper on \code{case_when} that associates FRAM FisheryID values according to CRC catch area codes.
A character-class "area_code" field in the input object is required and must have values with a minimum of 2 characters.
For example, an "areacode" of 1 should be "01", via something like \code{area_code = stringr::str_pad(areacode, width = 2, pad = "0")}
}
